update- enum Role moved to user/domain
